requirements.txt
Flask
beautifulsoup4
requests
SQLAlchemy
transformers
feedparser
APScheduler==3.10.1
pytz
python-dotenv  # For managing environment variables
